Log internal errors instead of forwarding them to the user
In case they happen the user can't do anything about them anyway, since
they are unexpected, so no need to create an extra ID for them.

Also don't exposed internal error messages to the user as they could contain
sensitive information.

Instead log the exception away, so it can be used for debugging,
and just give a generic error message to the user.