Add an additional audit logger to the payment service and log the demo mode
Up until now the main mono bundle doesn't use the audit logging channel.
Add one and log the demo mode message there since it might be relevant to
to the process.

I couldn't find a nice (bundle local) way to add a second logger with a different
channel to a serive, so this creates a dummy service that gets tagged with
a monolog channel, so the channel gets created, and then the channel is also
injected into the real service based on the arguments.