cleanup: handle prepared payments properly
In case a payment was created but not started it is in the
state "prepared" and doesn't have a payment method associated with
it. Up until now it would try to find the method, and the contract
and try to clean it up via the service provider connector and
fail because everything is null there.

In case there is no payment method just delete the payment as is,
since no other related data exists.

In case the method is set, but the contract can't be found this means
the config might have changed in the meantime. Since this means we
can't clean up everything, just give up and log an error, so the
entires can be manually cleaned up, or the config fixed.