backend interface: don't update the payment data after it was started
Once a payment is started we don't want to change the payment data in
case something changes in the backend system, as that is in most cases
only an error, and makes it hard to track what exactly was sent to
the payment provider.

This leads to the problem that we currently depend on the backend updating
the alternateName of the payment with translations and not doing so once
the payment is started would break the translation support in the API responses.

To work around this introduce a new method that takes a (read-only) DB entry
and the payment entity before it is sent to the API user. This allows the
backend to set the translated text at runtime in all cases.

While we are at it changing the interface also rename the parameters
and don't pass the payment objects by reference, since there is no
reason why the backend is allowed to replace them.