Remove the cancelled payment status
It is still unused, so remove it for now.

It's not clear if there is any implementation where we could know
the difference between cancelled and failed and even if we would,
we would probably show the same result to the user in both cases,
so just try to use failed instead.

If there is a valid use case for cancelled we can bring it back.