When trying to complete a payment, ignore the session timeout
We currently set a session timeout when the user creates a payment
and also when the suers starts is (defaults to 30min). After that
time has passed the user can no longer access the payment via the API.

The problem is that if the complete callback comes 30 minutes after
the payment has started we would try to fetch the payment via the
same path and fail.

The payment is current fetch in three places, when fetching it via the
API, when starting it, and when completing it. In the two first cases
we already check the session timeout anyway and return a proper API
error, and in the last case we don't want to check the session timeout
since we want to complete payments that take longer. So just remove the
"timeout" check in "findOneActive()" alltogether, to fix this.