Restore the timeout check when completing the payment via the API
completePayAction() was used internally for the webhook as well as
by the API. Removing the timeout check also removed it for the
API users, while in theory it doesn't hurt there, or at least doesn't
make things worse, as fetching the payment status will fail too.

Let's split things up and have an internal and external way to
complete a payment for now.