cleanup: add --force-unknown option to clean entires with dangling references
In case the payment entry doesn't find the corresponding types in the config
it will be skipped by default. By passing "--force-unknown" it will just
skip the connectors it doesn't find in the config and clean up the rest.

This is useful if you remove a connector from the config and want to remove
the left over entries from the DB with having the connector loaded.