cleanup: introduce a default cleanup timeout config + rename config
Rename "timeout_before" to "retention_period", so it's easier to understand.

Add a default "retention_period" under "default_retention_period", which is
used for all statuses that are not explicitely configured. Defaults to null,
so defaults to no cleanup.

To make it possible to have a default and still skip cleanup for
some statuses allow setting the "retention_period" of a status to null, to
override the default.