payment status: allow read-only access on the payment in case we are in a read-only state
In case the session has expired, but we are in a state where the user can no longer change
the payment we can show the status of the payment to the user.

This allows to make the session timeout smaller without stripping users of the ability of
checking the payment status (for example if the payment processor takes a long time).